mac
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@132 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_core.c
@@ -536,7 +536,7 @@ SWITCH_DECLARE(switch_status) switch_core_session_read_frame(switch_core_session
 		switch_frame *enc_frame, *read_frame = *frame;
 
 		if (read_frame->codec) {
-			size_t flag = 0;
+			unsigned int flag = 0;
 			session->raw_read_frame.datalen = session->raw_read_frame.buflen;
 			status = switch_core_codec_decode(read_frame->codec,
 										   session->read_codec,
@@ -637,7 +637,7 @@ SWITCH_DECLARE(switch_status) switch_core_session_write_frame(switch_core_sessio
 
 	switch_status status = SWITCH_STATUS_FALSE;
 	switch_frame *enc_frame, *write_frame = frame;
-	int flag = 0, need_codec = 0, perfect = 0;
+	unsigned int flag = 0, need_codec = 0, perfect = 0;
 	switch_io_flag io_flag = SWITCH_IO_FLAG_NOOP;
